Migrate skeleton as a more traditional web app (#6)
Migrate the skeleton angular app to a traditional web server app that will be run as a local server and interfaced with a small electron app.

https://github.com/ild-games/Skein/issues/7

Originally, the plan was to create an electron based angular application, but now the decision is to move to a traditional web server application that is spawned via an electron application and then communicated with via Get/Post requests to local web server. This is explained more in https://github.com/ild-games/Skein/issues/8 .